create execution mining and state computation sub module for state manager
