+++ src/state_manager/mod.rs
@@ -10,8 +10,11 @@ mod cache;
 pub mod chain_rand;
 pub mod circulating_supply;
 mod errors;
+mod execution;
 mod message_search;
 mod message_simulation;
+mod mining;
+mod state_computation;
 pub mod utils;
 
 pub use self::errors::*;
@@ -205,6 +208,14 @@ pub struct StateManager<DB> {
 
 #[allow(clippy::type_complexity)]
 pub const NO_CALLBACK: Option<fn(MessageCallbackCtx<'_>) -> anyhow::Result<()>> = None;
+
+/// Controls whether the VM should flush its state after execution
+#[derive(Debug, Copy, Clone, Default)]
+pub enum VMFlush {
+    Flush,
+    #[default]
+    Skip,
+}
 
 impl<DB> StateManager<DB>
 where
@@ -448,1028 +459,4 @@ where
         let state = miner::State::load(self.blockstore(), actor.code, actor.state)?;
         state.load_sectors_ext(self.blockstore(), None)
     }
-}
-
-impl<DB> StateManager<DB>
-where
-    DB: Blockstore + Send + Sync + 'static,
-{
-    /// Load the state of a tipset, including state root, message receipts
-    pub async fn load_tipset_state(self: &Arc<Self>, ts: &Tipset) -> anyhow::Result<TipsetState> {
-        if let Some(state) = self.cache.get_map(ts.key(), |et| et.into()) {
-            Ok(state)
-        } else {
-            match self.chain_store().load_child_tipset(ts)? {
-                Some(receipt_ts) => Ok(TipsetState {
-                    state_root: *receipt_ts.parent_state(),
-                    receipt_root: *receipt_ts.parent_message_receipts(),
-                }),
-                None => Ok(self.load_executed_tipset(ts).await?.into()),
-            }
-        }
-    }
-
-    /// Load an executed tipset, including state root, message receipts and events with caching.
-    pub async fn load_executed_tipset(
-        self: &Arc<Self>,
-        ts: &Tipset,
-    ) -> anyhow::Result<ExecutedTipset> {
-        // validate the existence of state trees for post-chain-head-epoch tipsets in case chain head is reset(e.g. manually or via GC).
-        if ts.epoch() >= self.heaviest_tipset().epoch()
-            && let Some(cached) = self.cache.get(ts.key())
-        {
-            if StateTree::new_from_root(self.blockstore_owned(), &cached.state_root).is_ok() {
-                return Ok(cached);
-            } else {
-                self.cache.remove(ts.key());
-            }
-        }
-        self.cache
-            .get_or_else(ts.key(), || async move {
-                let receipt_ts = self.chain_store().load_child_tipset(ts)?;
-                self.load_executed_tipset_inner(ts, receipt_ts.as_ref())
-                    .await
-            })
-            .await
-    }
-
-    async fn load_executed_tipset_inner(
-        self: &Arc<Self>,
-        msg_ts: &Tipset,
-        // when `msg_ts` is the current head, `receipt_ts` is `None`
-        receipt_ts: Option<&Tipset>,
-    ) -> anyhow::Result<ExecutedTipset> {
-        if let Some(receipt_ts) = receipt_ts {
-            anyhow::ensure!(
-                msg_ts.key() == receipt_ts.parents(),
-                "message tipset should be the parent of message receipt tipset"
-            );
-        }
-        let mut recomputed = false;
-        let (state_root, receipt_root, receipts) = match receipt_ts.and_then(|ts| {
-            let receipt_root = *ts.parent_message_receipts();
-            Receipt::get_receipts(self.cs.blockstore(), receipt_root)
-                .ok()
-                .map(|r| (*ts.parent_state(), receipt_root, r))
-        }) {
-            Some((state_root, receipt_root, receipts)) => (state_root, receipt_root, receipts),
-            None => {
-                let state_output = self
-                    .compute_tipset_state(msg_ts.shallow_clone(), NO_CALLBACK, VMTrace::NotTraced)
-                    .await?;
-                recomputed = true;
-                (
-                    state_output.state_root,
-                    state_output.receipt_root,
-                    Receipt::get_receipts(self.cs.blockstore(), state_output.receipt_root)?,
-                )
-            }
-        };
-
-        let messages = self.chain_store().messages_for_tipset(msg_ts)?;
-        anyhow::ensure!(
-            messages.len() == receipts.len(),
-            "mismatching message and receipt counts ({} messages, {} receipts)",
-            messages.len(),
-            receipts.len()
-        );
-        let mut executed_messages = Vec::with_capacity(messages.len());
-        for (message, receipt) in messages.iter().cloned().zip(receipts) {
-            let events = if let Some(events_root) = receipt.events_root() {
-                Some(
-                    match StampedEvent::get_events(self.cs.blockstore(), &events_root) {
-                        Ok(events) => events,
-                        Err(e) if recomputed => return Err(e),
-                        Err(_) => {
-                            self.compute_tipset_state(
-                                msg_ts.shallow_clone(),
-                                NO_CALLBACK,
-                                VMTrace::NotTraced,
-                            )
-                            .await?;
-                            recomputed = true;
-                            StampedEvent::get_events(self.cs.blockstore(), &events_root)?
-                        }
-                    },
-                )
-            } else {
-                None
-            };
-            executed_messages.push(ExecutedMessage {
-                message,
-                receipt,
-                events,
-            });
-        }
-        Ok(ExecutedTipset {
-            state_root,
-            receipt_root,
-            executed_messages: Arc::new(executed_messages),
-        })
-    }
-
-    /// Replays the given message and returns the result of executing the
-    /// indicated message, assuming it was executed in the indicated tipset.
-    pub async fn replay(self: &Arc<Self>, ts: Tipset, mcid: Cid) -> Result<ApiInvocResult, Error> {
-        let this = Arc::clone(self);
-        tokio::task::spawn_blocking(move || this.replay_blocking(ts, mcid)).await?
-    }
-
-    /// Blocking version of `replay`
-    pub fn replay_blocking(
-        self: &Arc<Self>,
-        ts: Tipset,
-        mcid: Cid,
-    ) -> Result<ApiInvocResult, Error> {
-        const REPLAY_HALT: &str = "replay_halt";
-
-        let mut api_invoc_result = None;
-        let callback = |ctx: MessageCallbackCtx<'_>| {
-            match ctx.at {
-                CalledAt::Applied | CalledAt::Reward
-                    if api_invoc_result.is_none() && ctx.cid == mcid =>
-                {
-                    api_invoc_result = Some(ApiInvocResult {
-                        msg_cid: ctx.message.cid(),
-                        msg: ctx.message.message().clone(),
-                        msg_rct: Some(ctx.apply_ret.msg_receipt()),
-                        error: ctx.apply_ret.failure_info().unwrap_or_default(),
-                        duration: ctx.duration.as_nanos().clamp(0, u128::from(u64::MAX)) as u64,
-                        gas_cost: MessageGasCost::new(ctx.message.message(), ctx.apply_ret)?,
-                        execution_trace: structured::parse_events(ctx.apply_ret.exec_trace())
-                            .unwrap_or_default(),
-                    });
-                    anyhow::bail!(REPLAY_HALT);
-                }
-                _ => Ok(()), // ignored
-            }
-        };
-        let result = self.compute_tipset_state_blocking(ts, Some(callback), VMTrace::Traced);
-        if let Err(error_message) = result
-            && error_message.to_string() != REPLAY_HALT
-        {
-            return Err(Error::Other(format!(
-                "unexpected error during execution : {error_message:}"
-            )));
-        }
-        api_invoc_result.ok_or_else(|| Error::Other("failed to replay".into()))
-    }
-
-    /// Replays a tipset up to a target message, capturing the state root before
-    /// and after execution.
-    pub async fn replay_for_prestate(
-        self: &Arc<Self>,
-        ts: Tipset,
-        target_message_cid: Cid,
-    ) -> Result<(Cid, ApiInvocResult, Cid), Error> {
-        let this = Arc::clone(self);
-        tokio::task::spawn_blocking(move || {
-            this.replay_for_prestate_blocking(ts, target_message_cid)
-        })
-        .await
-        .map_err(|e| Error::Other(format!("{e}")))?
-    }
-
-    fn replay_for_prestate_blocking(
-        self: &Arc<Self>,
-        ts: Tipset,
-        target_msg_cid: Cid,
-    ) -> Result<(Cid, ApiInvocResult, Cid), Error> {
-        if ts.epoch() == 0 {
-            return Err(Error::Other(
-                "cannot trace messages in the genesis block".into(),
-            ));
-        }
-
-        let genesis_timestamp = self.chain_store().genesis_block_header().timestamp;
-        let exec = TipsetExecutor::new(
-            self.chain_index().shallow_clone(),
-            self.chain_config().shallow_clone(),
-            self.beacon_schedule().shallow_clone(),
-            &self.engine,
-            ts.shallow_clone(),
-        );
-        let mut no_cb = NO_CALLBACK;
-        let (parent_state, epoch, block_messages) =
-            exec.prepare_parent_state(genesis_timestamp, VMTrace::NotTraced, &mut no_cb)?;
-
-        Ok(stacker::grow(64 << 20, || {
-            let mut vm =
-                exec.create_vm(parent_state, epoch, ts.min_timestamp(), VMTrace::NotTraced)?;
-            let mut processed = ahash::HashSet::default();
-
-            for block in block_messages.iter() {
-                let mut penalty = TokenAmount::zero();
-                let mut gas_reward = TokenAmount::zero();
-
-                for msg in block.messages.iter() {
-                    let cid = msg.cid();
-                    if processed.contains(&cid) {
-                        continue;
-                    }
-
-                    processed.insert(cid);
-
-                    if cid == target_msg_cid {
-                        let pre_root = vm.flush()?;
-                        let mut traced_vm =
-                            exec.create_vm(pre_root, epoch, ts.min_timestamp(), VMTrace::Traced)?;
-                        let (ret, duration) = traced_vm.apply_message(msg)?;
-                        let post_root = traced_vm.flush()?;
-
-                        return Ok((
-                            pre_root,
-                            ApiInvocResult {
-                                msg_cid: cid,
-                                msg: msg.message().clone(),
-                                msg_rct: Some(ret.msg_receipt()),
-                                error: ret.failure_info().unwrap_or_default(),
-                                duration: duration.as_nanos().clamp(0, u128::from(u64::MAX)) as u64,
-                                gas_cost: MessageGasCost::default(),
-                                execution_trace: structured::parse_events(ret.exec_trace())
-                                    .unwrap_or_default(),
-                            },
-                            post_root,
-                        ));
-                    }
-
-                    let (ret, _) = vm.apply_message(msg)?;
-                    gas_reward += ret.miner_tip();
-                    penalty += ret.penalty();
-                }
-
-                if let Some(rew_msg) =
-                    vm.reward_message(epoch, block.miner, block.win_count, penalty, gas_reward)?
-                {
-                    let (ret, _) = vm.apply_implicit_message(&rew_msg)?;
-                    if let Some(err) = ret.failure_info() {
-                        bail!(
-                            "failed to apply reward message for miner {}: {err}",
-                            block.miner
-                        );
-                    }
-
-                    // This is more of a sanity check, this should not be able to be hit.
-                    if !ret.msg_receipt().exit_code().is_success() {
-                        bail!(
-                            "reward application message failed (exit: {:?})",
-                            ret.msg_receipt().exit_code()
-                        );
-                    }
-                }
-            }
-
-            bail!("message {target_msg_cid} not found in tipset")
-        })?)
-    }
-
-    /// Checks the eligibility of the miner. This is used in the validation that
-    /// a block's miner has the requirements to mine a block.
-    pub fn eligible_to_mine(
-        &self,
-        address: &Address,
-        base_tipset: &Tipset,
-        lookback_tipset: &Tipset,
-    ) -> anyhow::Result<bool, Error> {
-        let hmp =
-            self.miner_has_min_power(&self.chain_config().policy, address, lookback_tipset)?;
-        let version = self.get_network_version(base_tipset.epoch());
-
-        if version <= NetworkVersion::V3 {
-            return Ok(hmp);
-        }
-
-        if !hmp {
-            return Ok(false);
-        }
-
-        let actor = self
-            .get_actor(&Address::POWER_ACTOR, *base_tipset.parent_state())?
-            .ok_or_else(|| Error::state("Power actor address could not be resolved"))?;
-
-        let power_state = power::State::load(self.blockstore(), actor.code, actor.state)?;
-
-        let actor = self
-            .get_actor(address, *base_tipset.parent_state())?
-            .ok_or_else(|| Error::state("Miner actor address could not be resolved"))?;
-
-        let miner_state = miner::State::load(self.blockstore(), actor.code, actor.state)?;
-
-        // Non-empty power claim.
-        let claim = power_state
-            .miner_power(self.blockstore(), address)?
-            .ok_or_else(|| Error::Other("Could not get claim".to_string()))?;
-        if claim.quality_adj_power <= BigInt::zero() {
-            return Ok(false);
-        }
-
-        // No fee debt.
-        if !miner_state.fee_debt().is_zero() {
-            return Ok(false);
-        }
-
-        // No active consensus faults.
-        let info = miner_state.info(self.blockstore())?;
-        if base_tipset.epoch() <= info.consensus_fault_elapsed {
-            return Ok(false);
-        }
-
-        Ok(true)
-    }
-
-    /// Conceptually, a [`Tipset`] consists of _blocks_ which share an _epoch_.
-    /// Each _block_ contains _messages_, which are executed by the _Filecoin Virtual Machine_.
-    ///
-    /// VM message execution essentially looks like this:
-    /// ```text
-    /// state[N-900..N] * message = state[N+1]
-    /// ```
-    ///
-    /// The `state`s above are stored in the `IPLD Blockstore`, and can be referred to by
-    /// a [`Cid`] - the _state root_.
-    /// The previous 900 states (configurable, see
-    /// <https://docs.filecoin.io/reference/general/glossary/#finality>) can be
-    /// queried when executing a message, so a store needs at least that many.
-    /// (a snapshot typically contains 2000, for example).
-    ///
-    /// Each message costs FIL to execute - this is _gas_.
-    /// After execution, the message has a _receipt_, showing how much gas was spent.
-    /// This is similarly a [`Cid`] into the block store.
-    ///
-    /// For details, see the documentation for [`apply_block_messages`].
-    ///
-    pub async fn compute_tipset_state(
-        self: &Arc<Self>,
-        tipset: Tipset,
-        callback: Option<impl FnMut(MessageCallbackCtx<'_>) -> anyhow::Result<()> + Send + 'static>,
-        enable_tracing: VMTrace,
-    ) -> Result<ExecutedTipset, Error> {
-        let this = Arc::clone(self);
-        tokio::task::spawn_blocking(move || {
-            this.compute_tipset_state_blocking(tipset, callback, enable_tracing)
-        })
-        .await?
-    }
-
-    /// Blocking version of `compute_tipset_state`
-    pub fn compute_tipset_state_blocking(
-        &self,
-        tipset: Tipset,
-        callback: Option<impl FnMut(MessageCallbackCtx<'_>) -> anyhow::Result<()>>,
-        enable_tracing: VMTrace,
-    ) -> Result<ExecutedTipset, Error> {
-        let epoch = tipset.epoch();
-        let has_callback = callback.is_some();
-        info!(
-            "Evaluating tipset: EPOCH={epoch}, blocks={}, tsk={}",
-            tipset.len(),
-            tipset.key(),
-        );
-        Ok(apply_block_messages(
-            self.chain_store().genesis_block_header().timestamp,
-            self.chain_index().shallow_clone(),
-            self.chain_config().shallow_clone(),
-            self.beacon_schedule().shallow_clone(),
-            &self.engine,
-            tipset,
-            callback,
-            enable_tracing,
-        )
-        .map_err(|e| {
-            if has_callback {
-                e
-            } else {
-                e.context(format!("Failed to compute tipset state@{epoch}"))
-            }
-        })?)
-    }
-
-    #[instrument(skip_all)]
-    pub async fn compute_state(
-        self: &Arc<Self>,
-        height: ChainEpoch,
-        messages: Vec<Message>,
-        tipset: Tipset,
-        callback: Option<impl FnMut(MessageCallbackCtx<'_>) -> anyhow::Result<()> + Send + 'static>,
-        enable_tracing: VMTrace,
-    ) -> Result<ExecutedTipset, Error> {
-        let this = Arc::clone(self);
-        tokio::task::spawn_blocking(move || {
-            this.compute_state_blocking(height, messages, tipset, callback, enable_tracing)
-        })
-        .await?
-    }
-
-    /// Blocking version of `compute_state`
-    #[tracing::instrument(skip_all)]
-    pub fn compute_state_blocking(
-        &self,
-        height: ChainEpoch,
-        messages: Vec<Message>,
-        tipset: Tipset,
-        callback: Option<impl FnMut(MessageCallbackCtx<'_>) -> anyhow::Result<()>>,
-        enable_tracing: VMTrace,
-    ) -> Result<ExecutedTipset, Error> {
-        Ok(compute_state(
-            height,
-            messages,
-            tipset,
-            self.chain_store().genesis_block_header().timestamp,
-            self.chain_index().shallow_clone(),
-            self.chain_config().shallow_clone(),
-            self.beacon_schedule().shallow_clone(),
-            &self.engine,
-            callback,
-            enable_tracing,
-        )?)
-    }
-
-    pub async fn miner_get_base_info(
-        self: &Arc<Self>,
-        beacon_schedule: &BeaconSchedule,
-        tipset: Tipset,
-        addr: Address,
-        epoch: ChainEpoch,
-    ) -> anyhow::Result<Option<MiningBaseInfo>> {
-        let prev_beacon = self
-            .chain_store()
-            .chain_index()
-            .latest_beacon_entry(tipset.clone())?;
-
-        let entries: Vec<BeaconEntry> = beacon_schedule
-            .beacon_entries_for_block(
-                self.chain_config().network_version(epoch),
-                epoch,
-                tipset.epoch(),
-                &prev_beacon,
-            )
-            .await?;
-
-        let base = entries.last().unwrap_or(&prev_beacon);
-
-        let (lb_tipset, lb_state_root) = ChainStore::get_lookback_tipset_for_round(
-            self.chain_index(),
-            self.chain_config(),
-            &tipset,
-            epoch,
-        )?;
-
-        // If the miner actor doesn't exist in the current tipset, it is a
-        // user-error and we must return an error message. If the miner exists
-        // in the current tipset but not in the lookback tipset, we may not
-        // error and should instead return None.
-        let actor = self.get_required_actor(&addr, *tipset.parent_state())?;
-        if self.get_actor(&addr, lb_state_root)?.is_none() {
-            return Ok(None);
-        }
-
-        let miner_state = miner::State::load(self.blockstore(), actor.code, actor.state)?;
-
-        let addr_buf = to_vec(&addr)?;
-        let rand = draw_randomness(
-            base.signature(),
-            DomainSeparationTag::WinningPoStChallengeSeed as i64,
-            epoch,
-            &addr_buf,
-        )?;
-
-        let network_version = self.chain_config().network_version(tipset.epoch());
-        let sectors = self.get_sectors_for_winning_post(
-            &lb_state_root,
-            network_version,
-            &addr,
-            Randomness::new(rand.to_vec()),
-        )?;
-
-        if sectors.is_empty() {
-            return Ok(None);
-        }
-
-        let (miner_power, total_power) = self
-            .get_power(&lb_state_root, Some(&addr))?
-            .context("failed to get power")?;
-
-        let info = miner_state.info(self.blockstore())?;
-
-        let worker_key = self
-            .resolve_to_deterministic_address(info.worker, &tipset)
-            .await?;
-        let eligible = self.eligible_to_mine(&addr, &tipset, &lb_tipset)?;
-
-        Ok(Some(MiningBaseInfo {
-            miner_power: miner_power.quality_adj_power,
-            network_power: total_power.quality_adj_power,
-            sectors,
-            worker_key,
-            sector_size: info.sector_size,
-            prev_beacon_entry: prev_beacon,
-            beacon_entries: entries,
-            eligible_for_mining: eligible,
-        }))
-    }
-
-    /// Checks power actor state for if miner meets consensus minimum
-    /// requirements.
-    pub fn miner_has_min_power(
-        &self,
-        policy: &Policy,
-        addr: &Address,
-        ts: &Tipset,
-    ) -> anyhow::Result<bool> {
-        let actor = self
-            .get_actor(&Address::POWER_ACTOR, *ts.parent_state())?
-            .ok_or_else(|| Error::state("Power actor address could not be resolved"))?;
-        let ps = power::State::load(self.blockstore(), actor.code, actor.state)?;
-
-        ps.miner_nominal_power_meets_consensus_minimum(policy, self.blockstore(), addr)
-    }
-
-    /// Validates all tipsets at epoch `start..=end` behind the heaviest tipset.
-    ///
-    /// Tipsets are processed sequentially. The compute-intensive work inside each
-    /// tipset (`bellperson` proof verification, FVM batch seal verification, etc.)
-    /// is already heavily rayon-parallelized. Parallelizing the outer loop actually introduces
-    /// some issues due to locks in the aforementioned crates. So don't do it.
-    ///
-    /// # What is validation?
-    /// Every state transition returns a new _state root_, which is typically retained in, e.g., snapshots.
-    /// For "full" snapshots, all state roots are retained.
-    /// For standard snapshots, the last 2000 or so state roots are retained.
-    ///
-    /// _receipts_ meanwhile, are typically ephemeral, but each tipset knows the _receipt root_
-    /// (hash) of the previous tipset.
-    ///
-    /// This function takes advantage of that fact to validate tipsets:
-    /// - `tipset[N]` claims that `receipt_root[N-1]` should be `0xDEADBEEF`
-    /// - find `tipset[N-1]`, and perform its state transition to get the actual `receipt_root`
-    /// - assert that they match
-    ///
-    /// See [`Self::compute_tipset_state_blocking`] for an explanation of state transitions.
-    #[tracing::instrument(skip(self))]
-    pub fn validate_range(&self, epochs: RangeInclusive<i64>) -> anyhow::Result<()> {
-        let heaviest = self.heaviest_tipset();
-        let heaviest_epoch = heaviest.epoch();
-        let end = self.chain_index().load_required_tipset_by_height(
-            *epochs.end(),
-            heaviest,
-            ResolveNullTipset::TakeOlder,
-        ).with_context(|| {
-            format!(
-        "couldn't get a tipset at height {} behind heaviest tipset at height {heaviest_epoch}",
-        *epochs.end(),
-    )})?;
-
-        // lookup tipset parents as we go along, iterating DOWN from `end`
-        let tipsets = end
-            .chain(self.blockstore())
-            .take_while(|ts| ts.epoch() >= *epochs.start());
-
-        self.validate_tipsets(tipsets)
-    }
-
-    pub fn validate_tipsets<T>(&self, tipsets: T) -> anyhow::Result<()>
-    where
-        T: Iterator<Item = Tipset> + Send,
-    {
-        let genesis_timestamp = self.chain_store().genesis_block_header().timestamp;
-        validate_tipsets(
-            genesis_timestamp,
-            self.chain_index(),
-            self.chain_config(),
-            self.beacon_schedule(),
-            &self.engine,
-            tipsets,
-        )
-    }
-
-    pub fn execution_trace(&self, tipset: &Tipset) -> anyhow::Result<(Cid, Vec<ApiInvocResult>)> {
-        let mut invoc_trace = vec![];
-
-        let genesis_timestamp = self.chain_store().genesis_block_header().timestamp;
-
-        let callback = |ctx: MessageCallbackCtx<'_>| {
-            match ctx.at {
-                CalledAt::Applied | CalledAt::Reward => {
-                    invoc_trace.push(ApiInvocResult {
-                        msg_cid: ctx.message.cid(),
-                        msg: ctx.message.message().clone(),
-                        msg_rct: Some(ctx.apply_ret.msg_receipt()),
-                        error: ctx.apply_ret.failure_info().unwrap_or_default(),
-                        duration: ctx.duration.as_nanos().clamp(0, u128::from(u64::MAX)) as u64,
-                        gas_cost: MessageGasCost::new(ctx.message.message(), ctx.apply_ret)?,
-                        execution_trace: structured::parse_events(ctx.apply_ret.exec_trace())
-                            .unwrap_or_default(),
-                    });
-                    Ok(())
-                }
-                _ => Ok(()), // ignored
-            }
-        };
-
-        let ExecutedTipset { state_root, .. } = apply_block_messages(
-            genesis_timestamp,
-            self.chain_index().shallow_clone(),
-            self.chain_config().shallow_clone(),
-            self.beacon_schedule().shallow_clone(),
-            &self.engine,
-            tipset.shallow_clone(),
-            Some(callback),
-            VMTrace::Traced,
-        )?;
-
-        Ok((state_root, invoc_trace))
-    }
-}
-
-pub fn validate_tipsets<DB, T>(
-    genesis_timestamp: u64,
-    chain_index: &ChainIndex<DB>,
-    chain_config: &Arc<ChainConfig>,
-    beacon: &Arc<BeaconSchedule>,
-    engine: &MultiEngine,
-    tipsets: T,
-) -> anyhow::Result<()>
-where
-    DB: Blockstore + Send + Sync + 'static,
-    T: Iterator<Item = Tipset> + Send,
-{
-    // Validate one tipset at a time. Parallelizing the outer loop across tipsets
-    // might wedge the global rayon pool.
-    // Sequential outer iteration leaves the entire rayon pool free for that
-    // already-rich inner parallelism.
-    for (child, parent) in tipsets.tuple_windows() {
-        info!(height = parent.epoch(), "compute parent state");
-        let ExecutedTipset {
-            state_root: actual_state,
-            receipt_root: actual_receipt,
-            ..
-        } = apply_block_messages(
-            genesis_timestamp,
-            chain_index.shallow_clone(),
-            chain_config.shallow_clone(),
-            beacon.shallow_clone(),
-            engine,
-            parent,
-            NO_CALLBACK,
-            VMTrace::NotTraced,
-        )
-        .context("couldn't compute tipset state")?;
-        let expected_receipt = child.min_ticket_block().message_receipts;
-        let expected_state = child.parent_state();
-        if (expected_state, expected_receipt) != (&actual_state, actual_receipt) {
-            error!(
-                height = child.epoch(),
-                ?expected_state,
-                ?expected_receipt,
-                ?actual_state,
-                ?actual_receipt,
-                "state mismatch"
-            );
-            bail!("state mismatch");
-        }
-    }
-    Ok(())
-}
-
-/// Shared context for creating VMs and preparing tipset state.
-///
-/// Encapsulates randomness source, genesis info, VM construction,
-/// null-epoch cron handling, and state migrations.
-struct TipsetExecutor<'a, DB: Blockstore + Send + Sync + 'static> {
-    tipset: Tipset,
-    rand: ChainRand<DB>,
-    chain_config: Arc<ChainConfig>,
-    chain_index: ChainIndex<DB>,
-    genesis_info: GenesisInfo,
-    engine: &'a MultiEngine,
-}
-
-impl<'a, DB: Blockstore + Send + Sync + 'static> TipsetExecutor<'a, DB> {
-    fn new(
-        chain_index: ChainIndex<DB>,
-        chain_config: Arc<ChainConfig>,
-        beacon: Arc<BeaconSchedule>,
-        engine: &'a MultiEngine,
-        tipset: Tipset,
-    ) -> Self {
-        let rand = ChainRand::new(
-            chain_config.shallow_clone(),
-            tipset.shallow_clone(),
-            chain_index.shallow_clone(),
-            beacon,
-        );
-        let genesis_info = GenesisInfo::from_chain_config(chain_config.shallow_clone());
-        Self {
-            tipset,
-            rand,
-            chain_config,
-            chain_index,
-            genesis_info,
-            engine,
-        }
-    }
-
-    fn create_vm(
-        &self,
-        state_root: Cid,
-        epoch: ChainEpoch,
-        timestamp: u64,
-        trace: VMTrace,
-    ) -> anyhow::Result<VM<DB>> {
-        let circ_supply = self.genesis_info.get_vm_circulating_supply(
-            epoch,
-            self.chain_index.db(),
-            &state_root,
-        )?;
-        VM::new(
-            ExecutionContext {
-                heaviest_tipset: self.tipset.shallow_clone(),
-                state_tree_root: state_root,
-                epoch,
-                rand: Box::new(self.rand.shallow_clone()),
-                base_fee: self.tipset.min_ticket_block().parent_base_fee.clone(),
-                circ_supply,
-                chain_config: self.chain_config.shallow_clone(),
-                chain_index: self.chain_index.shallow_clone(),
-                timestamp,
-            },
-            self.engine,
-            trace,
-        )
-    }
-
-    /// Produces the state root ready for message execution by running
-    /// null-epoch `crons` and any pending state migrations.
-    fn prepare_parent_state<F>(
-        &self,
-        genesis_timestamp: u64,
-        null_epoch_trace: VMTrace,
-        cron_callback: &mut Option<F>,
-    ) -> anyhow::Result<(Cid, ChainEpoch, Vec<BlockMessages>)>
-    where
-        F: FnMut(MessageCallbackCtx<'_>) -> anyhow::Result<()>,
-    {
-        use crate::shim::clock::EPOCH_DURATION_SECONDS;
-
-        let mut parent_state = *self.tipset.parent_state();
-        let parent_epoch = self
-            .chain_index
-            .load_required_tipset(self.tipset.parents())?
-            .epoch();
-        let epoch = self.tipset.epoch();
-
-        for epoch_i in parent_epoch..epoch {
-            if epoch_i > parent_epoch {
-                let timestamp = genesis_timestamp + ((EPOCH_DURATION_SECONDS * epoch_i) as u64);
-                parent_state = stacker::grow(64 << 20, || -> anyhow::Result<Cid> {
-                    let mut vm =
-                        self.create_vm(parent_state, epoch_i, timestamp, null_epoch_trace)?;
-                    if let Err(e) = vm.run_cron(epoch_i, cron_callback.as_mut()) {
-                        error!("Beginning of epoch cron failed to run: {e:#}");
-                        return Err(e);
-                    }
-                    vm.flush()
-                })?;
-            }
-            if let Some(new_state) = run_state_migrations(
-                epoch_i,
-                &self.chain_config,
-                self.chain_index.db(),
-                &parent_state,
-            )? {
-                parent_state = new_state;
-            }
-        }
-
-        let block_messages = BlockMessages::for_tipset(self.chain_index.db(), &self.tipset)?;
-        Ok((parent_state, epoch, block_messages))
-    }
-}
-
-/// Messages are transactions that produce new states. The state (usually
-/// referred to as the 'state-tree') is a mapping from actor addresses to actor
-/// states. Each block contains the hash of the state-tree that should be used
-/// as the starting state when executing the block messages.
-///
-/// # Execution environment
-///
-/// Transaction execution has the following inputs:
-/// - a current state-tree (stored as IPLD in a key-value database). This
-///   reference is in [`Tipset::parent_state`].
-/// - up to 900 past state-trees. See
-///   <https://docs.filecoin.io/reference/general/glossary/#finality>.
-/// - up to 900 past tipset IDs.
-/// - a deterministic source of randomness.
-/// - the circulating supply of FIL (see
-///   <https://filecoin.io/blog/filecoin-circulating-supply/>). The circulating
-///   supply is determined by the epoch and the states of a few key actors.
-/// - the base fee (see <https://spec.filecoin.io/systems/filecoin_vm/gas_fee/>).
-///   This value is defined by `tipset.parent_base_fee`.
-/// - the genesis timestamp (UNIX epoch time when the first block was
-///   mined/created).
-/// - a chain configuration (maps epoch to network version, has chain specific
-///   settings).
-///
-/// The result of running a set of block messages is an index to the final
-/// state-tree and an index to an array of message receipts (listing gas used,
-/// return codes, etc).
-///
-/// # Cron and null tipsets
-///
-/// Once per epoch, after all messages have run, a special 'cron' transaction
-/// must be executed. The tasks of the 'cron' transaction include running batch
-/// jobs and keeping the state up-to-date with the current epoch.
-///
-/// It can happen that no blocks are mined in an epoch. The tipset for such an
-/// epoch is called a null tipset. A null tipset has no identity and cannot be
-/// directly executed. This is a problem for 'cron' which must run for every
-/// epoch, even if there are no messages. The fix is to run 'cron' if there are
-/// any null tipsets between the current epoch and the parent epoch.
-///
-/// Imagine the blockchain looks like this with a null tipset at epoch 9:
-///
-/// ```text
-/// ┌────────┐ ┌────┐ ┌───────┐  ┌───────┐
-/// │Epoch 10│ │Null│ │Epoch 8├──►Epoch 7├─►
-/// └───┬────┘ └────┘ └───▲───┘  └───────┘
-///     └─────────────────┘
-/// ```
-///
-/// The parent of tipset-epoch-10 is tipset-epoch-8. Before executing the
-/// messages in epoch 10, we have to run cron for epoch 9. However, running
-/// 'cron' requires the timestamp of the youngest block in the tipset (which
-/// doesn't exist because there are no blocks in the tipset). Lotus dictates that
-/// the timestamp of a null tipset is `30s * epoch` after the genesis timestamp.
-/// So, in the above example, if the genesis block was mined at time `X`, the
-/// null tipset for epoch 9 will have timestamp `X + 30 * 9`.
-///
-/// # Migrations
-///
-/// Migrations happen between network upgrades and modify the state tree. If a
-/// migration is scheduled for epoch 10, it will be run _after_ the messages for
-/// epoch 10. The tipset for epoch 11 will link the state-tree produced by the
-/// migration.
-///
-/// Example timeline with a migration at epoch 10:
-///   1. Tipset-epoch-10 executes, producing state-tree A.
-///   2. Migration consumes state-tree A and produces state-tree B.
-///   3. Tipset-epoch-11 executes, consuming state-tree B (rather than A).
-///
-/// Note: The migration actually happens when tipset-epoch-11 executes. This is
-///       because tipset-epoch-10 may be null and therefore not executed at all.
-///
-/// # Caching
-///
-/// Scanning the blockchain to find past tipsets and state-trees may be slow.
-/// The `ChainStore` caches recent tipsets to make these scans faster.
-#[allow(clippy::too_many_arguments)]
-pub fn apply_block_messages<DB>(
-    genesis_timestamp: u64,
-    chain_index: ChainIndex<DB>,
-    chain_config: Arc<ChainConfig>,
-    beacon: Arc<BeaconSchedule>,
-    engine: &MultiEngine,
-    tipset: Tipset,
-    mut callback: Option<impl FnMut(MessageCallbackCtx<'_>) -> anyhow::Result<()>>,
-    enable_tracing: VMTrace,
-) -> anyhow::Result<ExecutedTipset>
-where
-    DB: Blockstore + Send + Sync + 'static,
-{
-    // This function will:
-    // 1. handle the genesis block as a special case
-    // 2. run 'cron' for any null-tipsets between the current tipset and our parent tipset
-    // 3. run migrations
-    // 4. execute block messages
-    // 5. write the state-tree to the DB and return the CID
-
-    // step 1: special case for genesis block
-    if tipset.epoch() == 0 {
-        // NB: This is here because the process that executes blocks requires that the
-        // block miner reference a valid miner in the state tree. Unless we create some
-        // magical genesis miner, this won't work properly, so we short circuit here
-        // This avoids the question of 'who gets paid the genesis block reward'
-        let message_receipts = tipset.min_ticket_block().message_receipts;
-        return Ok(ExecutedTipset {
-            state_root: *tipset.parent_state(),
-            receipt_root: message_receipts,
-            executed_messages: vec![].into(),
-        });
-    }
-
-    let exec = TipsetExecutor::new(
-        chain_index.shallow_clone(),
-        chain_config,
-        beacon,
-        engine,
-        tipset.shallow_clone(),
-    );
-
-    // step 2: running cron for any null-tipsets
-    // step 3: run migrations
-    let (parent_state, epoch, block_messages) =
-        exec.prepare_parent_state(genesis_timestamp, enable_tracing, &mut callback)?;
-
-    // FVM requires a stack size of 64MiB. The alternative is to use `ThreadedExecutor` from
-    // FVM, but that introduces some constraints, and possible deadlocks.
-    stacker::grow(64 << 20, || -> anyhow::Result<ExecutedTipset> {
-        let mut vm = exec.create_vm(parent_state, epoch, tipset.min_timestamp(), enable_tracing)?;
-
-        // step 4: apply tipset messages
-        let (receipts, events, events_roots) =
-            vm.apply_block_messages(&block_messages, epoch, callback)?;
-
-        // step 5: construct receipt root from receipts
-        let receipt_root = Amtv0::new_from_iter(chain_index.db(), receipts.iter())?;
-
-        // step 6: store events AMTs in the blockstore
-        for (events, events_root) in events.iter().zip(events_roots.iter()) {
-            if let Some(events) = events {
-                let event_root =
-                    events_root.context("events root should be present when events present")?;
-                // Store the events AMT - the root CID should match the one computed by FVM
-                let derived_event_root = Amt::new_from_iter_with_bit_width(
-                    chain_index.db(),
-                    EVENTS_AMT_BITWIDTH,
-                    events.iter(),
-                )
-                .map_err(|e| Error::Other(format!("failed to store events AMT: {e}")))?;
-
-                // Verify the stored root matches the FVM-computed root
-                ensure!(
-                    derived_event_root == event_root,
-                    "Events AMT root mismatch: derived={derived_event_root}, actual={event_root}."
-                );
-            }
-        }
-
-        let state_root = vm.flush()?;
-
-        // Update executed tipset cache
-        let messages: Vec<ChainMessage> = block_messages
-            .into_iter()
-            .flat_map(|bm| bm.messages)
-            .collect_vec();
-        anyhow::ensure!(
-            messages.len() == receipts.len() && messages.len() == events.len(),
-            "length of messages, receipts, and events should match",
-        );
-        Ok(ExecutedTipset {
-            state_root,
-            receipt_root,
-            executed_messages: messages
-                .into_iter()
-                .zip(receipts)
-                .zip(events)
-                .map(|((message, receipt), events)| ExecutedMessage {
-                    message,
-                    receipt,
-                    events,
-                })
-                .collect_vec()
-                .into(),
-        })
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn compute_state<DB>(
-    _height: ChainEpoch,
-    messages: Vec<Message>,
-    tipset: Tipset,
-    genesis_timestamp: u64,
-    chain_index: ChainIndex<DB>,
-    chain_config: Arc<ChainConfig>,
-    beacon: Arc<BeaconSchedule>,
-    engine: &MultiEngine,
-    callback: Option<impl FnMut(MessageCallbackCtx<'_>) -> anyhow::Result<()>>,
-    enable_tracing: VMTrace,
-) -> anyhow::Result<ExecutedTipset>
-where
-    DB: Blockstore + Send + Sync + 'static,
-{
-    if !messages.is_empty() {
-        anyhow::bail!("Applying messages is not yet implemented.");
-    }
-
-    let output = apply_block_messages(
-        genesis_timestamp,
-        chain_index,
-        chain_config,
-        beacon,
-        engine,
-        tipset,
-        callback,
-        enable_tracing,
-    )?;
-
-    Ok(output)
-}
-
-/// Controls whether the VM should flush its state after execution
-#[derive(Debug, Copy, Clone, Default)]
-pub enum VMFlush {
-    Flush,
-    #[default]
-    Skip,
 }
